# hands-on-work-vii/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API de Painéis Gráficos de Gestão Imobiliária - Hands on Work VII
  description: >
    Serviços REST para alimentar os painéis gráficos (barras, linha/dispersão e
    pizza) de uma empresa de gestão imobiliária. Os dados são carregados na
    íntegra a partir do banco relacional e filtrados/agregados na aplicação,
    por meio de técnicas de programação funcional (map/filter/reduce), não no
    SGBD. Especificação dos 3 serviços implementados na Parte 2.
  version: 1.0.0
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Ambiente local de desenvolvimento
security: []

paths:
  /api/imoveis/total-pagamentos:
    get:
      summary: Total acumulado de pagamentos por imóvel
      description: >
        Gráfico de barras: para cada imóvel, o somatório de todos os
        pagamentos ao longo de toda a série histórica.
      operationId: getTotalPagamentosPorImovel
      responses:
        '200':
          description: Lista com o total acumulado de pagamentos por imóvel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TotalPorImovel'
              example:
                - codigoImovel: 3
                  totalPagamentos: 30000.00
                - codigoImovel: 9
                  totalPagamentos: 12000.00
        '500':
          $ref: '#/components/responses/ErroInterno'

  /api/vendas/mensal:
    get:
      summary: Total de vendas por mês/ano
      description: >
        Gráfico de linha/dispersão: total de vendas ocorridas em cada
        mês/ano, considerando todos os dados disponíveis.
      operationId: getVendasPorMes
      responses:
        '200':
          description: Lista com o total de vendas por mês/ano
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TotalPorMes'
              example:
                - mesAno: '11/2025'
                  totalVendas: 23000.00
                - mesAno: '12/2025'
                  totalVendas: 15000.00
        '500':
          $ref: '#/components/responses/ErroInterno'

  /api/imoveis/percentual-por-tipo:
    get:
      summary: Percentual do valor total de vendas por tipo de imóvel
      description: >
        Gráfico de pizza: percentual do valor total de vendas de cada
        modalidade de imóvel (Apartamento, Casa, Terreno, Sala Comercial)
        em relação a toda a série histórica.
      operationId: getPercentualVendasPorTipoImovel
      responses:
        '200':
          description: Lista com o percentual de vendas por tipo de imóvel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PercentualPorTipo'
              example:
                - tipoImovel: Apartamento
                  percentual: 50.0
                - tipoImovel: Casa
                  percentual: 30.0
                - tipoImovel: Sala Comercial
                  percentual: 20.0
        '500':
          $ref: '#/components/responses/ErroInterno'

components:
  responses:
    ErroInterno:
      description: Erro inesperado ao processar a série histórica
      content:
        application/json:
          schema:
            type: object
            properties:
              mensagem:
                type: string
                example: Falha ao consultar pagamentos
  schemas:
    TotalPorImovel:
      type: object
      properties:
        codigoImovel:
          type: integer
          example: 3
        totalPagamentos:
          type: number
          format: double
          example: 30000.00
    TotalPorMes:
      type: object
      properties:
        mesAno:
          type: string
          description: Mês e ano no formato MM/AAAA
          example: '11/2025'
        totalVendas:
          type: number
          format: double
          example: 23000.00
    PercentualPorTipo:
      type: object
      properties:
        tipoImovel:
          type: string
          example: Apartamento
        percentual:
          type: number
          format: double
          example: 50.0